requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
optimum-intel==1.27.0
openvino==2026.0.0
huggingface-hub==0.36.2
Pillow==12.1.1
numpy==2.2.5
pytest==8.4.1
